added support for rack & pinion Z probe (Z_RACK_PINION)
Added support for compiling with Z_RACK_PINION for machines like Wilson2, which have a Z probe mounted to a pinion that is deployed by the movement of the X axis.
Changed the auto bedleveling default limits to be based on the bed size instead of absolute values (configuration.h).